@@ -12,11 +12,11 @@ classifiers =
     Operating System :: OS Independent
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
     Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: Implementation :: PyPy
 keywords = xml, html, diff
@@ -28,7 +28,7 @@ project_urls =
     Source Code = https://github.com/Shoobx/xmldiff
 
 [options]
-python_requires = >=3.7
+python_requires = >=3.8
 zip_safe = True
 include_package_data = True
 packages = find:
