@@ -1,11 +1,11 @@
 """All major API points and command-line tools"""
-import pkg_resources
+from importlib import metadata
 
 from argparse import ArgumentParser, ArgumentTypeError
 from lxml import etree
 from xmldiff import diff, formatting, patch
 
-__version__ = pkg_resources.require("xmldiff")[0].version
+__version__ = metadata.version("xmldiff")
 
 FORMATTERS = {
     "diff": formatting.DiffFormatter,
